requirements.txt
streamlit==1.28.0
snowflake-connector-python==3.5.0
snowflake-snowpark-python==1.11.0
python-dotenv==1.0.0
pydantic==2.5.0

